pyproject.toml
[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "x10-python-trading-starknet"
version = "1.0.5"
description = "Python client for X10 API"
authors = ["X10 <tech@ex10.org>"]
repository = "https://github.com/x10xchange/python_sdk"
documentation = "https://api.docs.extended.exchange/"
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{ include = "x10" }, { include = "extended" }]

[tool.poetry.dependencies]
aiohttp = ">=3.10.11"
eth-account = ">=0.12.0"
fast-stark-crypto = "==0.3.8"
pydantic = ">=2.9.0"
python = "^3.10"
pyyaml = ">=6.0.1"
sortedcontainers = ">=2.4.0"
strenum = "^0.4.15"
tenacity = "^9.1.2"
nest-asyncio = ">=1.5.6"
websockets = "^10.4"
propcache = ">=0.4.1"

[tool.poetry.group.dev.dependencies]
black = "==23.12.0"
flake8 = "==6.1.0"
flake8-bugbear = "==24.12.12"
freezegun = "==1.5.5"
isort = "==5.13.2"
mypy = "==1.18.2"
mypy-extensions = "==1.0.0"
PyHamcrest = "2.1.0"
pytest = "==7.4.3"
pytest-aiohttp = "==1.0.5"
pytest-asyncio = "==0.23.3"
pytest-cov = "==4.1.0"
pytest-forked = "==1.6.0"
pytest-mock = "==3.12.0"
python-dotenv = "==1.0.1"
safety = "==3.5.1"
tox = "==4.11.4"
types-pyyaml = "==6.0.12.12"
typing-extensions = ">=4.9.0"


[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--import-mode=importlib"

[tool.mypy]
packages = ["examples", "tests", "x10", "extended"]
plugins = ["pydantic.mypy"]
follow_untyped_imports = true
check_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
